+++ CHANGELOG.md
@@ -21,6 +21,9 @@
 - STAC browser now opens in a new tab.
 - Use react-oidc-context for authentication.
 - Use OpenID Connect "sub" claim as user identifier.
+- Expanded API health checks.
+- Authenticated users are now registed in the database.
+- First registered user is granted administrator privileges.
 
 ## v1.0.93 - 2024-05-14